Change mean shift parameter for number of iterations from int to size_t

+++ include/in-house-tracking/iht.h
@@ -19,14 +19,14 @@ void IHT_calc3DByteDepthUniformHist_CV(const Mat& image, Mat& hist);
 
 void IHT_calc3DByteDepthBackProject(const uchar* image, const uchar* hist, uchar* result, size_t imgrows, size_t imgcols, size_t imgstep);
 
-void IHT_meanShift(const uchar* densityMap, int maprows, int mapcols, int mapstep, window* w, int iters);
+void IHT_meanShift(const uchar* densityMap, int maprows, int mapcols, int mapstep, window* w, size_t iters);
 void IHT_meanShift_CV(const Mat& densityMap, Rect& window, size_t iters);
 
 extern "C" {
     void IHT_calc3DByteDepthUniformHist_ASM(const uchar* imgdata, uchar* histdata, size_t imgrows, size_t imgcols, size_t imgstep);
     void IHT_calc3DByteDepthUniformHist_DISASM(const uchar* imgdata, uchar* histdata, size_t imgrows, size_t imgcols, size_t imgstep);
     void IHT_calc3DByteDepthBackProject_ASM(const uchar* image, const uchar* hist, uchar* result, size_t imgrows, size_t imgcols, size_t imgstep);
-    void IHT_meanShift_ASM(const uchar* densityMap, int maprows, int mapcols, int mapstep, window* w, int iters);
+    void IHT_meanShift_ASM(const uchar* densityMap, int maprows, int mapcols, int mapstep, window* w, size_t iters);
 }
 
 /* aux */
